Added new testing asset <character/steve> with model, animatons, rigidbody, and a ragdoll. I have begun implementation of ragdoll physics. The skinned to physics has been finished and Physical subclasses now use arrays to keep track of hitboxes and physics objects.
Major bug involving multiple instances of an animated model has been
found, fix that ASAP. Left TODO in Model.java. Next step in ragdoll is
some kind of system to set up IK in the .ragdoll tag and have the engine
interpret it in a non-retarded way. This will be interesting.